+++ Cron/SendMailToAdmin.php
@@ -8,6 +8,7 @@ use Magento\Customer\Model\ResourceModel\Visitor\CollectionFactory as VisitorCol
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\Mail\Template\TransportBuilder;
 use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
+use Magento\Framework\Translate\Inline\StateInterface;
 use Overdose\CustomerPasswordReHash\Model\PasswordVerifier;
 
 class SendMailToAdmin
@@ -42,7 +43,12 @@ class SendMailToAdmin
     /**
      * @var TransportBuilder
      */
-    protected  $transportBuilder;
+    protected $transportBuilder;
+
+    /**
+     * @var StateInterface
+     */
+    protected $inlineTranslation;
 
     /**
      * SendMailToAdmin constructor.
@@ -60,7 +66,8 @@ class SendMailToAdmin
         PasswordVerifier $passwordVerifier,
         VisitorCollectionFactory $visitorCollectionFactory,
         TimezoneInterface $_localeDate,
-        TransportBuilder $transportBuilder
+        TransportBuilder $transportBuilder,
+        StateInterface $inlineTranslation
     ) {
         $this->scopeConfig = $scopeConfig;
         $this->customerCollection = $customerCollection;
@@ -68,6 +75,7 @@ class SendMailToAdmin
         $this->visitorCollectionFactory = $visitorCollectionFactory;
         $this->_localeDate = $_localeDate;
         $this->transportBuilder = $transportBuilder;
+        $this->inlineTranslation = $inlineTranslation;
     }
 
     /**
@@ -89,25 +97,11 @@ class SendMailToAdmin
         foreach ($customerCollection as $customer) {
             if ($customer->getPasswordHash()) {
                 $validatedHash = $this->validateHash($customer->getPasswordHash());
-                if ($validatedHash) {//if not rehashed check last visit
+                if (!$validatedHash) {//if not rehashed check last visit
                     $checked = $this->checkLastVizited($months, $customer->getId());
 
                     if ($checked) {
-                        $postObject = new \Magento\Framework\DataObject();
-                        $postObject->setData(['result'=> 'All passwords was rehashed.']);
-                        $emailTransportBuilder = $this->transportBuilder
-                        ->setTemplateIdentifier('od_send_message_admin_template')
-                        ->setTemplateOptions(
-                            [
-                                'area' => \Magento\Backend\App\Area\FrontNameResolver::AREA_CODE,
-                                'store' => \Magento\Store\Model\Store::DEFAULT_STORE_ID,
-                            ]
-                        )
-                        ->setTemplateVars(['data'=> $postObject])
-                        ->setFromByScope(['name' => 'All passwds are rehashed', 'email' => 'mykola.syniavskyi@overdose.digital'])
-                        ->addTo('mykola.syniavskyi@overdose.digital')
-                        ->getTransport();
-                        $emailTransportBuilder->sendMessage();
+                        $this->sentMail();
                     }
                 }
             }
@@ -115,6 +109,10 @@ class SendMailToAdmin
         return $this;
     }
 
+    /**
+     * @param $hash
+     * @return bool
+     */
     public function validateHash($hash)
     {
         if ($this->passwordVerifier->isBcrypt($hash) || $this->passwordVerifier->isSha256($hash)) {
@@ -123,6 +121,11 @@ class SendMailToAdmin
         return true;
     }
 
+    /**
+     * @param $months
+     * @param $customerId
+     * @return bool
+     */
     public function checkLastVizited($months, $customerId)
     {
         $currentTime = $this->_localeDate->date();
@@ -135,5 +138,44 @@ class SendMailToAdmin
         if (isset($lastVisitTime)) {
             return $modifiedTime > $lastVisitTime ? true : false;
         }
+    }
+
+    /**
+     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws \Magento\Framework\Exception\MailException
+     */
+    public function sentMail()
+    {
+        $this->inlineTranslation->suspend();
+        $sender = [
+            'name' => 'Overdose_CustomerPasswordReHash',
+            'email' => 'owner@example.com'
+        ];
+
+        $sentToEmail = $this->scopeConfig
+            ->getValue('trans_email/ident_general/email', \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+
+        $sentToName = $this->scopeConfig
+            ->getValue('trans_email/ident_general/name', \Magento\Store\Model\ScopeInterface::SCOPE_STORE);
+
+        $transport = $this->transportBuilder
+            ->setTemplateIdentifier('od_send_message_admin_template')
+            ->setTemplateOptions(
+                [
+                    'area' => \Magento\Backend\App\Area\FrontNameResolver::AREA_CODE,
+                    'store' => \Magento\Store\Model\Store::DEFAULT_STORE_ID,
+                ]
+            )
+            ->setTemplateVars([
+                'data'=> 'All passwords were rehashed! You can to disable module "Overdose_CustomerPasswordReHash".'
+            ])
+            ->setFromByScope($sender)
+            ->addTo($sentToEmail, $sentToName)
+            //->addTo('owner@example.com','owner')
+            ->getTransport();
+
+        $transport->sendMessage();
+
+        $this->inlineTranslation->resume();
     }
 }